+++ genport.h
@@ -106,7 +106,7 @@ struct genportContext_t : dbtool_t {
 				signatureCRC = crc32Name(signatureCRC, "P");
 			if (pSignature->flags & signature_t::SIGMASK_REQUIRED)
 				signatureCRC = crc32Name(signatureCRC, "R");
-			if (pSignature->flags & signature_t::SIGMASK_LOOKUP)
+			if (pSignature->flags & signature_t::SIGMASK_KEY)
 				signatureCRC = crc32Name(signatureCRC, "K");
 		}
 
@@ -328,7 +328,7 @@ struct genportContext_t : dbtool_t {
 				fprintf(f, "P");
 			if (pSignature->flags & signature_t::SIGMASK_REQUIRED)
 				fprintf(f, "R");
-			if (pSignature->flags & signature_t::SIGMASK_LOOKUP)
+			if (pSignature->flags & signature_t::SIGMASK_KEY)
 				fprintf(f, "K");
 
 			fprintf(f, "\"]\n");
@@ -413,7 +413,7 @@ struct genportContext_t : dbtool_t {
 				else if (*pFlags == 'R')
 					pSignature->flags |= signature_t::SIGMASK_REQUIRED;
 				else if (*pFlags == 'K')
-					pSignature->flags |= signature_t::SIGMASK_LOOKUP;
+					pSignature->flags |= signature_t::SIGMASK_KEY;
 				else
 					ctx.fatal("\n{\"error\":\"unknown flag\",\"where\":\"%s:%s:%d\",\"name\":\"%s\"}\n", __FUNCTION__, __FILE__, __LINE__, pName);