+++ CHANGELOG.md
@@ -10,6 +10,7 @@ TODO: no sorting when read-only
 ## [Unreleased]
 
 ```
+2021-08-09 15:39:31 Added: `genmember --cascade`.
 2021-08-03 13:34:46 Renamed: `signature_t::SIGMASK_LOCKED` to `signature_t::SIGMASK_KEY`.
 ```